chore(shared): reorganize UI folder structure
