+++ src/app/shared/ui/dialogs/confirm-dialog.component.scss
